% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalBayes-package.R
\docType{package}
\name{survivalBayes-package}
\alias{survivalBayes}
\alias{survivalBayes-package}
\title{survivalBayes: Bayesian Survival Analysis}
\description{
Implementation of Bayesian survival models with and without covariates. Optimized for common priors (normal, gamma, beta) and explores modern Bayesian inference techniques, such as Variational Bayes, for efficient computation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://victorney.github.io/survivalBayes/}
}

}
\author{
\strong{Maintainer}: Victor Ney \email{victor.s.ney@gmail.com}

}
\keyword{internal}
